requirements.txt
pandas
python-telegram-bot~=21.6
python-dotenv~=1.0.1
openai~=1.53.0